requirements.txt
customtkinter
pillow
qrcode
